Conflicts:
	smart-user-client-impl/src/main/java/com/smartitengineering/user/client/impl/App.java

+++ smart-user-resource/src/main/java/com/smartitengineering/user/ws/resources/OrganizationPrivilegeResource.java
@@ -34,22 +34,55 @@ import org.apache.commons.lang.StringUtils;
  *
  * @author russel
  */
-@Path("/orgs/sn/{organizationUniqueShortName}/privs/{privilegeName}")
+@Path("/orgs/sn/{organizationUniqueShortName}/privs/name/{privilegeName}")
 public class OrganizationPrivilegeResource extends AbstractResource {
 
+  static final UriBuilder PRIVILEGE_URI_BUILDER = UriBuilder.fromResource(OrganizationPrivilegeResource.class);
+  static final UriBuilder PRIVILEGE_CONTENT_URI_BUILDER;
+
+  static {
+    PRIVILEGE_CONTENT_URI_BUILDER = PRIVILEGE_URI_BUILDER.clone();
+    try {
+      PRIVILEGE_CONTENT_URI_BUILDER.path(OrganizationPrivilegeResource.class.getMethod("getContent"));
+    }
+    catch (Exception ex) {
+      ex.printStackTrace();
+      throw new InstantiationError();
+
+    }
+  }
   private String organizationUniqueShortName;
   private String privilegeName;
+  private Organization organization;
+  private Privilege privilege;
 
   public OrganizationPrivilegeResource(@PathParam("organizationUniqueShortName") String organizationUniqueShortName, @PathParam(
       "privilegeName") String privilegeName) {
     this.organizationUniqueShortName = organizationUniqueShortName;
     this.privilegeName = privilegeName;
+    organization = getOrganization();
+    privilege = getPrivilege();
+  }
+
+  @GET
+  @Produces(MediaType.APPLICATION_JSON)
+  @Path("/content")
+  public Response getContent() {
+    ResponseBuilder responseBuilder = Response.ok();
+    if (organization==null || privilege == null) {
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
+    responseBuilder = Response.ok(getPrivilege());
+    return responseBuilder.build();
   }
 
   @GET
   @Produces(MediaType.APPLICATION_ATOM_XML)
   public Response get() {
-    ResponseBuilder responseBuilder;
+    ResponseBuilder responseBuilder = Response.ok();
+    if (organization==null || privilege== null) {
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
     try {
       responseBuilder = Response.status(Status.OK);
       Feed privilegeFeed = getPrivilegeFeed();
@@ -64,7 +97,10 @@ public class OrganizationPrivilegeResource extends AbstractResource {
 
   @DELETE
   public Response delete() {
-    ResponseBuilder responseBuilder;
+    ResponseBuilder responseBuilder = Response.ok();
+    if (organization==null || privilege == null) {
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
     try {
       responseBuilder = Response.status(Status.OK);
       Services.getInstance().getPrivilegeService().delete(getPrivilege());
@@ -80,10 +116,13 @@ public class OrganizationPrivilegeResource extends AbstractResource {
   @Produces(MediaType.APPLICATION_ATOM_XML)
   @Consumes(MediaType.APPLICATION_JSON)
   public Response update(Privilege newPrivilege) {
-    ResponseBuilder responseBuilder;
+    ResponseBuilder responseBuilder = Response.ok();
+    if(organization==null || privilege==null){
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
     try {
       responseBuilder = Response.status(Status.OK);
-      newPrivilege.setParentOrganization(getOrganization());
+      newPrivilege.setParentOrganization(organization);
       Services.getInstance().getPrivilegeService().update(newPrivilege);
     }
     catch (Exception ex) {
@@ -108,13 +147,12 @@ public class OrganizationPrivilegeResource extends AbstractResource {
 
 
     Link altLink = abderaFactory.newLink();
-    altLink.setHref(UriBuilder.fromResource(OrganizationPrivilegeResource.class).build(organizationUniqueShortName,
-                                                                                       privilegeName).toString());
+    altLink.setHref(PRIVILEGE_CONTENT_URI_BUILDER.clone().build(organizationUniqueShortName, privilegeName).toString());
     altLink.setRel(Link.REL_ALTERNATE);
     altLink.setMimeType(MediaType.APPLICATION_JSON);
+    privilegeFeed.addLink(altLink);
 
     privilegeFeed.addLink(editLink);
-    privilegeFeed.addLink(altLink);
 
     return privilegeFeed;
   }
@@ -123,6 +161,9 @@ public class OrganizationPrivilegeResource extends AbstractResource {
   @Path("/delete")
   public Response deletePost() {
     ResponseBuilder responseBuilder = Response.ok();
+    if(organization==null || privilege==null){
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
     try {
       Services.getInstance().getPrivilegeService().delete(getPrivilege());
     }
@@ -138,6 +179,9 @@ public class OrganizationPrivilegeResource extends AbstractResource {
   @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
   public Response updatePost(@HeaderParam("Content-type") String contentType, String message) {
     ResponseBuilder responseBuilder = Response.status(Status.SERVICE_UNAVAILABLE);
+    if(organization==null || privilege==null){
+      return responseBuilder.status(Status.NOT_FOUND).build();
+    }
 
     if (StringUtils.isBlank(message)) {
       responseBuilder = Response.status(Status.BAD_REQUEST);
@@ -164,8 +208,7 @@ public class OrganizationPrivilegeResource extends AbstractResource {
         ex.printStackTrace();
       }
     }
-    else {
-      contentType = contentType;
+    else {      
       isHtmlPost = false;
     }
